src/backend/pyproject.toml
[project]
name = "hybrid-priority-engine"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "supabase",
    "google-genai",
    "httpx",
    "python-dotenv",
    "loguru",
    "tenacity",
    "python-multipart"
]
